Mod1Pr2
Write a program using a function and a series of if…else statements. The user should be prompted to enter a whole number in main in the range 1–5. The function is called and the number entered is passed to the function. This function parameter will then be used in a series of if…else statements to discover and print out the number word within the function (for example, 1=one, 2=two, 3=three, etc.).

Enter a number in the range 1–5: 1
You entered the number one.